refactor: renamed it to RasterClassifyLegend
