+++ sites/geohub/src/components/controls/RasterClassifyLegend.svelte
